pyproject.toml
[project]
name = "uetparser"
version = "0.1.0"
description = "ETL pipeline for uet"
authors = [
    {name = "Stepan",email = "s.ibatulov@yandex.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "googlesearch-python (>=1.3.0,<2.0.0)",
    "openai (>=1.93.0,<2.0.0)",
    "spacy (>=3.8.7,<4.0.0)",
    "scrapy (>=2.13.3,<3.0.0)",
    "html2text (>=2025.4.15,<2026.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "fuzzywuzzy (>=0.18.0,<0.19.0)",
    "python-levenshtein (>=0.27.1,<0.28.0)",
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "html-to-markdown (>=1.4.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "scrapy-user-agents (>=0.1.1,<0.2.0)",
    "scrapyrt (>=0.16.0,<0.17.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
